+++ src/test/inspection/scope/identifier.ts
@@ -3,9 +3,10 @@
 
 import "mocha";
 import { Inspection } from "../../..";
+import { ResultUtils } from "../../../common";
 import { Ast } from "../../../parser";
 import {
-    expectInspectionEqual,
+    expectDeepEqual,
     expectParseErrInspection,
     expectParseOkInspection,
     expectTextWithPosition,
@@ -13,7 +14,12 @@ import {
 
 type AbridgedScope = ReadonlyArray<string>;
 
-function actualFactoryFn(inspected: Inspection.Inspected): ReadonlyArray<string> {
+function actualFactoryFn(triedInspection: Inspection.TriedInspection): ReadonlyArray<string> {
+    if (!ResultUtils.isOk(triedInspection)) {
+        throw new Error(`AssertFailed: ResultUtils.isOk(triedInspection): ${triedInspection.error.message}`);
+    }
+    const inspected: Inspection.Inspected = triedInspection.value;
+
     return [...inspected.scope.keys()];
 }
 
@@ -22,31 +28,31 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`|each 1`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`|each 1`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`each| 1`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`each| 1`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`each |1`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`each |1`);
             const expected: AbridgedScope = ["_"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`each 1|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`each 1|`);
             const expected: AbridgedScope = ["_"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`each each 1|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`each each 1|`);
             const expected: AbridgedScope = ["_"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -54,13 +60,13 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`each|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`each|`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`each |`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`each |`);
             const expected: AbridgedScope = ["_"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -68,25 +74,25 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`|(x) => z`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`|(x) => z`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`(x|, y) => z`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`(x|, y) => z`);
             const expected: AbridgedScope = ["x"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`(x, y)| => z`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`(x, y)| => z`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`(x, y) => z|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`(x, y) => z|`);
             const expected: AbridgedScope = ["z", "x", "y"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -94,25 +100,25 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`|(x) =>`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`|(x) =>`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`(x|, y) =>`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`(x|, y) =>`);
             const expected: AbridgedScope = ["x"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`(x, y)| =>`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`(x, y)| =>`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`(x, y) =>|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`(x, y) =>|`);
             const expected: AbridgedScope = ["x", "y"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -120,43 +126,43 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`|foo`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`|foo`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`f|oo`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`f|oo`);
             const expected: AbridgedScope = ["foo"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`foo|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`foo|`);
             const expected: AbridgedScope = ["foo"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`|@foo`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`|@foo`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`@|foo`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`@|foo`);
             const expected: AbridgedScope = ["@foo"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`@f|oo`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`@f|oo`);
             const expected: AbridgedScope = ["@foo"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`@foo|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`@foo|`);
             const expected: AbridgedScope = ["@foo"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -164,31 +170,31 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`|foo(x)`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`|foo(x)`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`foo(x, y|)`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`foo(x, y|)`);
             const expected: AbridgedScope = ["y"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`foo(x, y)|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`foo(x, y)|`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[x](y|)`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[x](y|)`);
             const expected: AbridgedScope = ["y"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`foo(|)`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`foo(|)`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -196,37 +202,37 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`|foo(x`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`|foo(x`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`foo(x, y|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`foo(x, y|`);
             const expected: AbridgedScope = ["y"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`foo(|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`foo(|`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`foo(x,|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`foo(x,|`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`foo(x, |`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`foo(x, |`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[x](y|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[x](y|`);
             const expected: AbridgedScope = ["y"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -234,25 +240,25 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`|{1`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`|{1`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`{|1`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`{|1`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`{1|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`{1|`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`{{|1`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`{{|1`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -260,43 +266,43 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`|[a=1]`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`|[a=1]`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[|a=1]`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[|a=1]`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[a=1|]`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[a=1|]`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[a=1, b=2|]`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[a=1, b=2|]`);
             const expected: AbridgedScope = ["a"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[a=1, b=2|, c=3]`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[a=1, b=2|, c=3]`);
             const expected: AbridgedScope = ["a", "c"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[a=1]|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[a=1]|`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[a=[|b=1]]`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[a=[|b=1]]`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -304,49 +310,49 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`|[a=1`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`|[a=1`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[|a=1`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[|a=1`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[a=|1`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[a=|1`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[a=1|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[a=1|`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[a=1, b=|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[a=1, b=|`);
             const expected: AbridgedScope = ["a"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[a=1, b=2|, c=3`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[a=1, b=2|, c=3`);
             const expected: AbridgedScope = ["a", "c"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[a=[|b=1`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[a=[|b=1`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`[a=[b=|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`[a=[b=|`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -356,7 +362,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `s|ection foo; x = 1; y = 2;`,
             );
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`section foo; x = 1|; y = 2;`, () => {
@@ -364,7 +370,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `section foo; x = 1|; y = 2;`,
             );
             const expected: AbridgedScope = ["y"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`section foo; x = 1; y = 2|;`, () => {
@@ -372,7 +378,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `section foo; x = 1; y = 2|;`,
             );
             const expected: AbridgedScope = ["x"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`section foo; x = 1; y = 2;|`, () => {
@@ -380,7 +386,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `section foo; x = 1; y = 2;|`,
             );
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`section foo; x = 1; y = 2; z = let a = 1 in |b;`, () => {
@@ -388,7 +394,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `section foo; x = 1; y = 2; z = let a = 1 in |b;`,
             );
             const expected: AbridgedScope = ["a", "x", "y"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -398,7 +404,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `s|ection foo; x = 1; y = 2`,
             );
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`section foo; x = 1|; y = 2`, () => {
@@ -406,7 +412,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `section foo; x = 1|; y = 2`,
             );
             const expected: AbridgedScope = ["y"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`section foo; x = 1; y = 2|`, () => {
@@ -414,7 +420,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `section foo; x = 1; y = 2|`,
             );
             const expected: AbridgedScope = ["x"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`section foo; x = 1; y = () => 10|`, () => {
@@ -422,7 +428,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `section foo; x = 1; y = () => 10|`,
             );
             const expected: AbridgedScope = ["x"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -430,31 +436,31 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`let a = 1 in |x`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`let a = 1 in |x`);
             const expected: AbridgedScope = ["a"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`let a = 1 in x|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`let a = 1 in x|`);
             const expected: AbridgedScope = ["x", "a"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`let a = |1 in x`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`let a = |1 in x`);
             const expected: AbridgedScope = [];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`let a = 1, b = 2 in x|`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`let a = 1, b = 2 in x|`);
             const expected: AbridgedScope = ["x", "a", "b"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`let a = 1|, b = 2 in x`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`let a = 1|, b = 2 in x`);
             const expected: AbridgedScope = ["b"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`(p1, p2) => let a = 1, b = 2, c = 3| in c`, () => {
@@ -462,7 +468,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `(p1, p2) => let a = 1, b = 2, c = 3| in c`,
             );
             const expected: AbridgedScope = ["a", "b", "p1", "p2"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`let eggs = let ham = 0 in 1, foo = 2, bar = 3 in 4|`, () => {
@@ -470,7 +476,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `let eggs = let ham = 0 in 1, foo = 2, bar = 3 in 4|`,
             );
             const expected: AbridgedScope = ["eggs", "foo", "bar"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`let eggs = let ham = 0 in |1, foo = 2, bar = 3 in 4`, () => {
@@ -478,7 +484,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `let eggs = let ham = 0 in |1, foo = 2, bar = 3 in 4`,
             );
             const expected: AbridgedScope = ["ham", "foo", "bar"];
-            expectInspectionEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseOkInspection(text, position), expected, actualFactoryFn);
         });
     });
 
@@ -486,19 +492,19 @@ describe(`Inspection - Scope - Identifier`, () => {
         it(`let a = 1 in |`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`let a = 1 in |`);
             const expected: AbridgedScope = ["a"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`let a = 1, b = 2 in |`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`let a = 1, b = 2 in |`);
             const expected: AbridgedScope = ["a", "b"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`let a = 1|, b = 2 in`, () => {
             const [text, position]: [string, Inspection.Position] = expectTextWithPosition(`let a = 1|, b = 2 in `);
             const expected: AbridgedScope = ["b"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`let a = 1, b = 2, c = 3 in |`, () => {
@@ -506,7 +512,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `let a = 1, b = 2, c = 3 in |`,
             );
             const expected: AbridgedScope = ["a", "b", "c"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`let x = (let y = 1 in z|) in`, () => {
@@ -514,7 +520,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `let x = (let y = 1 in z|) in`,
             );
             const expected: AbridgedScope = ["z", "y"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
 
         it(`let x = (let y = 1 in z) in |`, () => {
@@ -522,7 +528,7 @@ describe(`Inspection - Scope - Identifier`, () => {
                 `let x = (let y = 1 in z) in |`,
             );
             const expected: AbridgedScope = ["x"];
-            expectInspectionEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
+            expectDeepEqual(expectParseErrInspection(text, position), expected, actualFactoryFn);
         });
     });
 });
